requirements.txt
flask
h5py
pillow
numpy
libmagic==0.4.6
